requirements.txt
psycopg2-binary
python-daemon
pika
mysql-connector
ftputil
requests
pymongo
redis
mock